refactor(cli): split main entrypoint helpers
