+++ docs/architecture/maintainability-baseline.txt
@@ -2,6 +2,5 @@
 # Format: <max_lines> <path>
 813 crates/surge-cli/src/commands/install/mod.rs
 1860 crates/surge-cli/src/commands/install/remote.rs
-835 crates/surge-cli/src/main.rs
 782 crates/surge-core/src/releases/delta.rs
 760 crates/surge-installer-ui/src/app.rs
